feat: enhance Vorpalstacks with tagging operations and expand SDK tests to 1387 cases
- Added TagResource and UntagResource operations for Timestream and Athena services.
- Updated README files to reflect the increase in passing Go SDK tests from 1311 to 1387.
- Improved timestamp formatting in various services to return float values for better precision.
- Implemented resource policy management for Kinesis streams, including PutResourcePolicy, GetResourcePolicy, and DeleteResourcePolicy operations.
- Expanded SDK tests to cover new tagging functionalities and ensure comprehensive service coverage.

+++ internal/services/aws/kinesis/consumer_operations.go
@@ -54,7 +54,7 @@ func (s *KinesisService) RegisterStreamConsumer(ctx context.Context, reqCtx *req
 			"ConsumerARN":               consumer.ConsumerARN,
 			"StreamARN":                 consumer.StreamARN,
 			"ConsumerStatus":            consumer.ConsumerStatus,
-			"ConsumerCreationTimestamp": consumer.ConsumerCreationTimestamp.Unix(),
+			"ConsumerCreationTimestamp": float64(consumer.ConsumerCreationTimestamp.Unix()),
 		},
 	}, nil
 }
@@ -127,7 +127,7 @@ func (s *KinesisService) DescribeStreamConsumer(ctx context.Context, reqCtx *req
 			"ConsumerARN":               consumer.ConsumerARN,
 			"StreamARN":                 consumer.StreamARN,
 			"ConsumerStatus":            consumer.ConsumerStatus,
-			"ConsumerCreationTimestamp": consumer.ConsumerCreationTimestamp.Unix(),
+			"ConsumerCreationTimestamp": float64(consumer.ConsumerCreationTimestamp.Unix()),
 		},
 	}, nil
 }
@@ -162,7 +162,7 @@ func (s *KinesisService) ListStreamConsumers(ctx context.Context, reqCtx *reques
 			"ConsumerARN":               c.ConsumerARN,
 			"StreamARN":                 c.StreamARN,
 			"ConsumerStatus":            c.ConsumerStatus,
-			"ConsumerCreationTimestamp": c.ConsumerCreationTimestamp.Unix(),
+			"ConsumerCreationTimestamp": float64(c.ConsumerCreationTimestamp.Unix()),
 		})
 	}
